pyproject.toml
[project]
name = "resume-screening"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.0"
dependencies = [
    "fastapi[standard]>=0.128.1",
    "openai>=2.17.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
